Delete dependencies.toml
Deleting Dependencies.toml due to lang level dependencies introduced through Ballerina 2201.4.x. Daily builds fail due to the changes (with nightly pack). 
Plan is to remove the file and let it build with latest libraries. At the time of releasing the connector with a released Ballerina version, plan is to take the Dependecies.toml generated with that distribution.